Add files via upload
All files are aptly named, if you are familiar with circuits they are fairly straightforward to use.

Some such as the Chapter 5 Q5, require key binds to use. This is two ram chips that should contain ARITH and LOGIC respectively, when the ALU switches between ARITH and LOGIC modes, this is displayed on the LCD. They are fairly intuitive and you can check which switches have certain keybinds.

You are more than welcome to contact me with any questions about them.